services/face_location/requirements.txt
certifi==2019.9.11
chardet==3.0.4
Click==7.0
face-recognition==1.2.3
face-recognition-models==0.3.0
idna==2.8
numpy==1.17.2
Pillow==6.1.0
redis==3.3.8
requests==2.22.0
urllib3==1.25.6
